requirements.txt
Flask
Flask-Mail
python-dotenv
Flask-Cors
gunicorn
